backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
scapy==2.5.0
pandas==2.1.3
joblib==1.3.2
scikit-learn==1.2.2
google-generativeai==0.3.1
pyjwt==2.8.0
cryptography==41.0.7
requests==2.31.0
tqdm
xgboost